Blenderbot/requirements.txt
streamlit==1.24.0
streamlit-chat==0.1.1
sentence-transformers==2.2.2

